requirements.txt
google-adk==1.1.1
feedparser==6.0.10
python-dateutil==2.8.2
python-dotenv==1.0.0 